requirements.txt
numpy
tensorflow
tensorflow_datasets
tensorrt
nvidia-cudnn-cu11
matplotlib
Pillow
nvidia-smi
keras_cv
PyYAML
pandas
accelerate
torchvision
transformers
datasets
ftfy
tensorboard
diffusers
xformers